requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
alive_progress==3.0.1
gradio==3.23.0
matplotlib==3.7.1
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
PyYAML==6.0
requests==2.28.2
setuptools==65.5.0
torch==2.0.0+cu117
tqdm==4.65.0
transformers==4.27.2
pyarrow==12.0.0
